requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.11.4
rerun-sdk==0.22.1
requests==2.31.0
